UPSTREAM: cpuidle: teo: Avoid code duplication in conditionals
There are three places in teo_select() where a given amount of time
is compared with TICK_NSEC if tick_nohz_tick_stopped() returns true,
which is a bit of duplicated code.

Avoid that code duplication by defining a helper function to do the
check and using it in all of the places in question.

No intentional functional impact.

Change-Id: I8236d066f3e6169daae1c4572b25b7396bb61a1c
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>